requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-dotenv==1.4.2
djangorestframework==3.12.2
ecdsa==0.14.1
gunicorn==20.0.4
oauthlib==3.1.0
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
python-jose==3.2.0
python3-openid==3.2.0
pytz==2021.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
social-auth-app-django==3.4.0
sqlparse==0.4.1
whitenoise==5.2.0
django-cors-headers~=3.1.1
drf-jwt~=1.13.3
pyjwt~=1.7.1
requests~=2.22.0